+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - CHANGELOG.md for tracking version history
+- Unit tests for intrinsics package (100% coverage)
 
 ## [1.0.0] - 2026-01-11